Attempt to fix issues with automatic migration being stuck in an infinite loop (#375)
* attempt to log more info

* more logging and make nothing actually do nothing

+++ src/core/user_settings.cpp
@@ -81,48 +81,51 @@ static void MovePath(fs::path const& _from, fs::path const& _to) {
 }
 
 static void CheckAndMigrateSaves(TransferOption option) {
+    LOG_INFO(Loader, "Starting save migration");
     auto const new_save_root = EmulatorSettings.GetHomeDir() / "1000" / "savedata";
     auto const old_save_root =
         Common::FS::GetUserPath(Common::FS::PathType::UserDir) / "savedata" / "1";
-    try {
-        for (const auto& entry : fs::directory_iterator(old_save_root)) {
-            if (!entry.is_directory()) {
-                continue;
-            }
-            const auto old_game_dir = entry.path();
-            const auto new_game_dir = new_save_root / old_game_dir.filename();
-            const bool already_exists = fs::exists(new_game_dir);
-
-            switch (option) {
-            case TransferOption::Copy:
-                if (!already_exists) {
-                    fs::copy(old_game_dir, new_game_dir, fs::copy_options::recursive);
-                }
-                break;
-            case TransferOption::Move:
-                if (!already_exists) {
-                    MovePath(old_game_dir, new_game_dir);
-                }
-                break;
-            case TransferOption::MoveAndLinkBack:
-                if (!already_exists) {
-                    MovePath(old_game_dir, new_game_dir);
-                    fs::create_directory_symlink(new_game_dir, old_game_dir);
-                }
-                break;
-            case TransferOption::Nothing:
-            case TransferOption::SdlCancelled:
-                return;
-            default:
-                UNREACHABLE();
-            }
-        }
-    } catch (std::exception const& e) {
-        UNREACHABLE_MSG("Error while migrating saves: {}", e.what());
+    if (!std::filesystem::exists(old_save_root)) {
+        return;
     }
+    for (const auto& entry : fs::directory_iterator(old_save_root)) {
+        if (!entry.is_directory()) {
+            continue;
+        }
+        const auto old_game_dir = entry.path();
+        const auto new_game_dir = new_save_root / old_game_dir.filename();
+        const bool already_exists = fs::exists(new_game_dir);
+        LOG_INFO(Loader, "Transferring {}", old_game_dir);
+
+        switch (option) {
+        case TransferOption::Copy:
+            if (!already_exists) {
+                fs::copy(old_game_dir, new_game_dir, fs::copy_options::recursive);
+            }
+            break;
+        case TransferOption::Move:
+            if (!already_exists) {
+                MovePath(old_game_dir, new_game_dir);
+            }
+            break;
+        case TransferOption::MoveAndLinkBack:
+            if (!already_exists) {
+                MovePath(old_game_dir, new_game_dir);
+                fs::create_directory_symlink(new_game_dir, old_game_dir);
+            }
+            break;
+        case TransferOption::Nothing:
+        case TransferOption::SdlCancelled:
+            return;
+        default:
+            UNREACHABLE();
+        }
+    }
+    LOG_INFO(Loader, "Save migration complete");
 }
 
 static void CheckAndMigrateTrophies(TransferOption option) {
+    LOG_INFO(Loader, "Starting trophy migration");
     const auto user_dir = EmulatorSettings.GetHomeDir() / "1000";
     const auto old_trophy_base_dir =
         Common::FS::GetUserPath(Common::FS::PathType::UserDir) / "game_data";
@@ -136,8 +139,9 @@ static void CheckAndMigrateTrophies(TransferOption option) {
         if (!entry.is_directory()) {
             continue;
         }
-
         const auto trophy_files_dir = entry.path() / "TrophyFiles";
+        LOG_INFO(Loader, "Transferring {}", trophy_files_dir);
+
         if (!fs::exists(trophy_files_dir)) {
             continue;
         }
@@ -148,6 +152,8 @@ static void CheckAndMigrateTrophies(TransferOption option) {
             }
 
             const auto old_trophy_dir = subentry.path();
+            LOG_INFO(Loader, "Transferring {}", old_trophy_dir);
+
             const auto xml_path = old_trophy_dir / "Xml" / "TROP.XML";
             if (!fs::exists(xml_path)) {
                 continue;
@@ -201,6 +207,7 @@ static void CheckAndMigrateTrophies(TransferOption option) {
             }
         }
     }
+    LOG_INFO(Loader, "Trophy migration complete");
 }
 
 void CheckSaveAndTrophyMigration() {
@@ -208,13 +215,22 @@ void CheckSaveAndTrophyMigration() {
         EmulatorSettings.GetHomeDir() / "1000" / "savedata" / ".data_transfer.complete";
     auto const old_save_dir =
         Common::FS::GetUserPath(Common::FS::PathType::UserDir) / "savedata" / "1";
-    if (fs::exists(old_save_dir) && !fs::is_empty(old_save_dir) &&
-        !fs::exists(migration_done_path)) {
-        TransferOption user_choice = AskMigrationOption();
-        CheckAndMigrateSaves(user_choice);
-        CheckAndMigrateTrophies(user_choice);
-        std::ofstream ofs(migration_done_path);
-        ofs << "";
+    try {
+        if (fs::exists(old_save_dir) && !fs::is_empty(old_save_dir) &&
+            !fs::exists(migration_done_path)) {
+            TransferOption user_choice = AskMigrationOption();
+            if (user_choice != TransferOption::Nothing &&
+                user_choice != TransferOption::SdlCancelled) {
+                CheckAndMigrateSaves(user_choice);
+                CheckAndMigrateTrophies(user_choice);
+            }
+            std::ofstream ofs(migration_done_path);
+            ofs << "";
+        }
+    } catch (std::exception const& e) {
+        QMessageBox::information(nullptr, "Save/Trophy migration",
+                                 QString("Error while transferring data: %1").arg(e.what()));
+        UNREACHABLE();
     }
 }
